Module gff_stats::stat

Statistical manipulations on a GFF3/fasta combination.

Structs

Output

Data structure sent through the parallel iterator.

SplicedCDS

Store the spliced CDS sequences

Writer

Wrapper for a vector of type Output.

Constants

HEADERS 🔒

The headers of the output table.

Functions

calculate_stats

Wrapper function to calculate statistics on sequences from a GFF3 file.

calculate_stats_spliced 🔒

Spliced version of the function which prints out statistics to STDOUT.

calculate_stats_unspliced 🔒

Non-spliced version of the function which prints out statistics to STDOUT.