Module gff_stats::utils

Utility functions used in other modules.

Structs

FourFoldStats

Statistics on the 4-fold-degenerate-sites.

GC3Stats

GC content at the third codon position.

Stats

Statistics on the whole sequence.

Constants

FOURFOLD_DEG

The fourfold degenerate codons.

SIXFOLD_DEG

The sixfold degenerate codons. Includes Leucine, Serine and Arginine.

Statics

AA_TABLE_CANONICAL

<https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi> U is equivalent to T here

ASCII_TO_INDEX

Maps an ASCII character to array index.

Functions

four_fold_site_stats

Take the vector of degenerate codons and calculate statistics.

gc_3

Calculate the GC content of the third position of a codon.

gc_four_fold_deg_sites

Take a trimmed sequence and return a vector of four/sixfold degenerate codons.

reverse_complement

Reverse complement a string of DNA.

switch_base

Switch bases to their complement.

translate

Taken directly from <https://github.com/dweb0/protein-translate/blob/master/src/lib.rs>

trim_sequence

From the fasta sequence iteration, get the subsequence and trim the sequence depending on the frame. If the sequence is not modulo 3, force it to be so.

whole_seq_stats

Simply calculate basic statistics on a &str.